backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv
google-cloud-storage
django-filter
cloud-sql-python-connector
psycopg2-binary
supabase
gotrue
python-jose 
pyjwt
celery
redis
django-celery-beat